Add files via upload
The objective of this project is to predict credit risk by conducting a bivariate data analysis to observe the impact of each variable on the target variable using the stacked bar plots. Various models are constructed based on the significance of the variables on the target variable, and the applicants’ credit risk is calculated.